incorporate metatheory extension (again)
