+++ Project.toml
@@ -12,6 +12,7 @@ TermInterface = "8ea1fca8-c5ef-4a55-8b96-4e9afe9c9a3c"
 [weakdeps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
+Metatheory = "e9d8d322-4543-424a-9be4-0cc815abe26c"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
@@ -19,6 +20,7 @@ SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 [extensions]
 SimpleExpressionsAbstractTreesExt = "AbstractTrees"
 SimpleExpressionsLatexifyExt = "Latexify"
+SimpleExpressionsMetatheoryExt = "Metatheory"
 SimpleExpressionsRecipesBaseExt = "RecipesBase"
 SimpleExpressionsRootsExt = "Roots"
 SimpleExpressionsSpecialFunctionsExt = "SpecialFunctions"
@@ -29,6 +31,7 @@ CallableExpressions = "1.0.0"
 Combinatorics = "1.0.2"
 CommonEq = "0.2.1"
 Latexify = "0.16, 1"
+Metatheory = "2, 3"
 RecipesBase = "1"
 Roots = "2"
 SpecialFunctions = "1,2"
@@ -37,8 +40,9 @@ julia = "1.10"
 
 [extras]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
+Metatheory = "e9d8d322-4543-424a-9be4-0cc815abe26c"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["Test", "Metatheory"]
