+++ test/runtests.jl
@@ -4,4 +4,4 @@ using Test
 import SimpleExpressions: @symbolic_expression
 
 include("basic_tests.jl")
-#include("extension_tests.jl")
+include("extension_tests.jl")
